refactor: Extract challenge translations to dedicated service
