+++ api/lib/infrastructure/transformers/challenge-transformer.js
@@ -1,13 +1,12 @@
 import _ from 'lodash';
-import { Challenge } from '../../domain/models/Challenge.js';
-import { fields as challengeLocalizedFields } from '../../infrastructure/translations/challenge.js';
+import { translateChallenges, translateChallenge } from '../../domain/services/translate-challenges.js';
 
 export function createChallengeTransformer({ attachments, localizedChallenges, localizedChallenge }) {
 
   if (localizedChallenges) {
     return _.flow(
       _addAttachmentsToChallenge({ attachments }),
-      _challengeToTranslatedChallenges({ localizedChallenges }),
+      translateChallenges({ localizedChallenges }),
       _filterChallengesFields,
     );
   }
@@ -15,7 +14,7 @@ export function createChallengeTransformer({ attachments, localizedChallenges, l
   if (localizedChallenge) {
     return _.flow(
       _addAttachmentsToChallenge({ attachments }),
-      (challenge) => _translateChallenge(challenge, localizedChallenge),
+      (challenge) => translateChallenge(challenge, localizedChallenge),
       _filterChallengeFields,
     );
   }
@@ -24,44 +23,6 @@ export function createChallengeTransformer({ attachments, localizedChallenges, l
     _addAttachmentsToChallenge({ attachments }),
     _filterChallengeFields,
   );
-}
-
-function _challengeToTranslatedChallenges({ localizedChallenges }) {
-  return (challenge) => localizedChallenges
-    .filter((localizedChallenge) => localizedChallenge.challengeId === challenge.id)
-    .map((localizedChallenge) => _translateChallenge(challenge, localizedChallenge));
-}
-
-function _translateChallenge(challenge, localizedChallenge) {
-  const primaryLocale = Challenge.getPrimaryLocale(challenge.locales) ?? 'fr';
-  const isPrimaryLocale = primaryLocale === localizedChallenge.locale;
-  const clearedLocalizedFields = challengeLocalizedFields.reduce((acc, field) => {
-    acc[field] = '';
-    return acc;
-  }, {});
-  return {
-    ...challenge,
-    ...clearedLocalizedFields,
-    ...challenge.translations[localizedChallenge.locale],
-    id: localizedChallenge.id,
-    status: isPrimaryLocale ? challenge.status : getLocalizedChallengeStatus(challenge, localizedChallenge),
-    embedUrl: localizedChallenge.embedUrl ?? _replaceLangParamsInUrl(localizedChallenge.locale, challenge.embedUrl),
-    locales: isPrimaryLocale ? challenge.locales : [localizedChallenge.locale],
-  };
-}
-
-function getLocalizedChallengeStatus(challenge, localizedChallenge) {
-  if (['proposé', 'périmé'].includes(challenge.status) || localizedChallenge.status === 'validé') {
-    return challenge.status;
-  }
-  return localizedChallenge.status;
-}
-
-function _replaceLangParamsInUrl(locale, embedUrl) {
-  if (!embedUrl) return undefined;
-  const url = new URL(embedUrl);
-  url.searchParams.set('lang', locale);
-  return url.href;
 }
 
 function _filterChallengesFields(challenges) {
